Starting on WB-626: Dynamic Correlation of Dense Time Series Without Computation of Dense Connectome: A preliminary implementation has been made.  The Jira ticked requested that the data-series file implementation be modified to support dynamic connectivity.  However, the matrix type files (dense) and map type file (data-series) are far too different to merge their functionality.  Some of the problems include one map versus many maps, yoking of map indices, identification, scenes, and so on.  While we could try merging these file types, it would (1) require months of effort; and (2), testing of existing functionality, especially scenes, would be an enormous challenge.  The design of these files took place at the start of the HCP and this type of mixed-functionality was not foreseen at that time.
The solution implemented adds a new CIFTI connectivity matrix file type called "dynamic dense" using the extension 'dynconn.nii'.  This file's internal format is identical to the data-series file format.  As a result, the user can copy a data-series file and change its filename extension and load it as a dynamic connectivity file.  This new file is a CIFTI connectivity matrix file so the user can select it as an overlay, click brainordinates, and the connectivity is computed by performing correlation operations.  The file can be saved to disk, saved to scenes, and the loaded data can be converted to a CIFTI Scalars File.