WB-729: Tab Dragging: Multiple signals (tabMoved(int,int) and currentChanged(int)) are emitted by QTabBar when a tab is moved.  The signal currentChanged(int) is also emitted when the user selects a tab.   The update graphics request has been removed from tabMoved() since it is not needed.  The timer logic for tab highlighting  in currentChanged() has been revised.  If the timer is running, it is canceled and restarted.  Previously, a static method in QTimer was used and it was possible for more than one of these timers to be active and causing graphics updates.
There is no signals emitted when moving of a tab starts and stops.  It might be possible to extend the QTabBar class and detect tab movement in the 'drag virtual methods' although they may not apply to a tab being moved.  Another possibility is to override the mouse press and release virtual methods.