Resolve WB-778 Scene does not reload a modified file: Each data file now contains the time that is was last read into memory or written out from memory.  When a scene is displayed, it avoid reloading in memory file that are unmodified.  This test has been expanded to compare the time associated with the file's last read or write action with the modified time on the disk file.  If the disk file time is newer, the file is reloaded.
A use case is as follows: The user is running wb_view and has an in memory file.  The user then uses wb_command or another instance of wb_view to change the file.  Now, when the user loads the scene, the new version of the file is loaded and reflects the content of the file on disk.