Implemented WB-780 Invert Palette Checkbox.  An "Invert" button has been added to the Palette tab of the Map Settings Dialog.  When selected, the palette is inverted (order of scalars/colors reversed and sign flipped).  An example usage is a positive only palette with data that is all negative.
Significant revisions were made to the way Palettes were accessed.  In many instances, code accessed the PaletteFile to get the palette using the name in the Palette Color Mapping.  All of this code has been removed and now the Palette is accessed through a method in PaletteColorMapping.  When the palette is requested, PaletteColorMapping will return the named Palette unless the "Inverted Palette" flag is set in which case an inverted version of the named Palette is returned.  As a result, code outside of PaletteColorMapping never needs to test the Inverted Flag (except for the dialog that turns the inverted flag on and off).

PaletteColorMapping obtains the Palette using EventPaletteGetByName instead of directly accessing the PaletteFile.  This may be helpful if user-defined palettes are ever added.

Lastly, if a palette name is invalid, the ROY-BIG-GL palette is used and a message is logged containing the invalid palette name.