Resolve WB-781 Apply to All Maps for ColorBar:  The "Apply to All Maps" was located on the Palette tab of the Overlay and Map Settings Dialog.  The request was for "Apply to All Maps" to affect the Color Bar tab of the dialog.  Since the Color Bar and Palette tabs were separate widgets, a new widget was created containing the "Apply to All Maps" user-interface controls.  This new widget is now used on both tabs.
A minor bug was found that if the user unchecked "Apply to All Maps" for a file but made no changes to palette settings, it could get set back on if the selected file in the map settings dialog was changed.  So, an "apply to all maps" property was added to CaretMappableDataFile.  This property is not saved to the file.  When the file is read, the property is initialized by verifying palette settings are identical in all maps.  This property is saved to and restored from scenes.

Some files (at least Data Series) use a "file palette" (on palette used for all maps within the file).  When this type of file is displayed in Map Settings Dialog, the "Apply to All Maps" checkbox is disabled as it is not needed with this type of file.