Implement WB-832 Graphics Region pop-up menu contains selections that "appear to do nothing":
This task is broken down into two issues.

(1) The menu contained every label from every label file (CIFTI or GIFTI) that was associated with the selected brainordinate.  So, if there were many label maps, the menu was long.  I have changed the logic to limit the labels to those displayed on the brain model in the selected tab (painted on the surface or volume).

(2) Three types of actions are available: (i) Show connectivity data from matrix type files; (ii) Show fiber trajectories; and (3) Show an average line chart.  The logic has been changed to so that these actions are only available when they are contained in an overlay from any tab.