WB-857: Older code fixed a problem with identical file names by saving the full path into the scene file and then performing a 'right-most characters" match to restore the data. While it works, we should not have absolute paths in the scene file.  It was probably done this way since the file was the name of a scene XML element and the scene path absolute/relative mechanism only works on 'ScenePathName' elements.
Now the name of the element is just the name of the file (no path).  A new child element, using a ScenePathName element has been added that is the absolute path in memory and a relative path in the scene file is now used to properly restore the data file's scene data.