Added a developer flag for a tile tabs layout modification as Develop Menu --> Tile Tabs Vertical Centering.  When this is enabled, each row in Tile Tabs is set to the minimum needed and the positions of the rows is adjusted so that any extra space is above all rows and below all rows; NOT between the rows.
A scene created in an older version of wb_view (September 2018) did not display correctly with extra space between the two rows of tabs.  The available height for each tab was greater than the height needed by aspect locking so tabs were taller than needed.  It is possible that the user locked aspect before entering tile tabs as this has been found to cause a similar layout problem.  Using the item in the Develop Menu, fixes this scene and others in the same scene file.