WB-859 Manual Tile Tabs Layout System: Added items to context menu when user right-clicks on a tab in the browser window and if a manual tab layout is currently selected: (1) Select for Manual Editing and (2) Show/Hide Tab Content in Window.  "Select for Manual Editing" is enable when the Mode is Tile and allows the user select the tab for editing is layout properties.  "Show/Hide Tab Content" is enabled when a manual layout is active and allows one to show or hide the tab's content in the graphics window.
Also, fixed bug in toolbar related to tab dragging that disable graphics/user-interface events from the time the mouse was pressed until the mouse was released.  However, when the pop-up (context sensitive) menu is displayed, the mouse pressed method is called by Qt but not the mouse released method.  So, when the mouse is pressed, only block signals if the the left-button is detected as down.