WB-926 Long scene descriptions cut off in scene selector: Added a command line option, "-scene-scroll   <number>" to wb_view that allows testing of two different options on the scene dialog for resolving this issue.  If this option is not specified, wb_view uses the existing functionality that chops off long scene descriptions.
"wb_view  -scene-scroll  1" places the description of the each scene inside a scroll area so that the user can scroll and read long scene descriptions.   The problem with this option is that it places "scroll regions" (for each scene) inside a larger scroll region (the scroll bar that scrolls all scenes). On a Mac (and maybe other platforms) this is cumbersome with trackpad scrolling.  If the mouse is over a scene description, the scene description is scrolled but if the user moves the mouse left of the scene description, the all scenes are scrolled.

 "wb_view  -scene-scroll  2" displays the full description for each scene.  If each scene contains a long description, the user will need to perform more scrolling to view each of the available scenes.  In my opinion, this is the preferred solution.

For testing a scene file that contains several scenes with long descriptions, try  /mnt/myelin/caret7_gui_design/data/HCP_WB_Tutorial_1.0/WB_1.0_SCENES.32k_fs_LR.scene.  Keep in mind this is an old scene file that contains "default descriptions" for each scene.  These "default descriptions" have been disabled for new scenes (but can be inserted using the "Add Window Info" button on the new scene dialog).