The frame-rate timer has been revised.  Previously, it was only measuring the time during OpenGL drawing (time spent in 'BrainOpenGLWidget::paintEvent') by starting and stopping the timer at the beginning and end of the function.  This timing excluded that spent swapping the frame buffer, performing user-interaction, user-interface updates, etc.  As a result, the frame rates were inaccurately too fast.
Now, the timing for the frame rate is updated once at the end of the graphics loop in 'BrainOpenGLWidget:paintGL'.  The 'frame-rate' is calculated from the entire time between each pair of successive calls that update the timer.  As a result, this interval of time includes buffer swapping, user-interface updates, user-interaction, etc and reflects the all of the time required to make one frame update.  One defect is that if the user does nothing for a while, the time between frames will include "that nothing".  To prevent this from distorting the frames-per-second, frames exceeding five seconds are excluded from any timing.  

The frames-per-second displayed in the window is an average of the time required to produce the last 10 frames.

Note that unlike a video game that may update its graphics at a regular interval, wb_view updates its graphics 'as needed'.